// types/screens-enum.ts
export enum SCREENS {
    SPLASH = "SPLASH",    
    ONBOARDING = "ONBOARDING",
    SIGNUP = "SIGNUP",
    EMAIL_VERIFICATION = "EMAIL_VERIFICATION",
    lOGIN = "lOGIN",
    FORGET_PASSWORD = "FORGET_PASSWORD",
    NEW_PASSWORD = "NEW_PASSWORD",
    NEW_PASSWORD_SUCCESS = "NEW_PASSWORD_SUCCESS",
    BOTTOM_NAV = "BOTTOM_NAV",
   




    // bottom
    HOME = "HOME",
    CATEGORIES="CATEGORIES",
    MYCART = "MYCART",
    WISHLIST="WISHLIST",
    PROFILE="PROFILE"


}